+++ projects/graphql-client/src/index.ts
@@ -10,3 +10,4 @@ export * from "./remote/queries/useSong"
 export * from "./remote/queries/useViewer"
 export * from "./remote/queries/useSongSearch"
 export * from "./remote/queries/useShares"
+export * from "./remote/queries/useShareUsers"
